refactor(web/stores): split workflows into package
Decomposes the 401-line stores/workflows.ts monolith into a
sibling-aggregator package. Pre-split: batchDeleteWorkflows cx 15
LPF 91, fetchMoreWorkflows cx 9. batchDeleteWorkflows decomposed
into settleBatchResults + pruneSucceeded + emitBatchToast + the
three buildXxxToast variants.

- stores/workflows.ts (38 lines): aggregator. updateFromWsEvent
  lives inline (one-line wrapper around fetchWorkflows).
- stores/workflows/types.ts: WorkflowsState plus BatchDeleteOutcome,
  WorkflowsSet, WorkflowsGet.
- stores/workflows/_state.ts: WORKFLOWS_PAGE_LIMIT + request token
  helpers (list + blueprint, both stale-checked).
- stores/workflows/list-actions.ts: fetchWorkflows, fetchMoreWorkflows
  (with a canFetchMore guard helper), loadBlueprints, setSearchQuery,
  setWorkflowTypeFilter.
- stores/workflows/crud-actions.ts: createWorkflow,
  createFromBlueprint, deleteWorkflow, batchDeleteWorkflows. The
  batch path is fully decomposed: settleBatchResults builds the
  succeeded/failed tallies, pruneSucceeded performs the state
  removal with concurrent-WS-aware recount, emitBatchToast picks
  one of three toast builders based on outcome.

Every new file under 400 LOC, every function under 80 LPF, every
function complexity at most 8. All 9 existing workflows tests pass.

Part of #2093 (PR B of EPIC #2066).
